requirements.txt
mlflow
gradio
google-generativeai
python-dotenv
pandas
scikit-learn
